modify: ログイン、ユーザー新規登録をusersに切り出し
