+++ app/basic_information/forms.py
@@ -1,12 +1,5 @@
 from django import forms
-from django.contrib.auth.forms import UserCreationForm
-from .models import CustomUser, Question
-
-
-class CustomUserCreationForm(UserCreationForm):
-    class Meta:
-        model = CustomUser
-        fields = ("username", "password1", "password2")  # 必要に応じてフィールド追加
+from .models import Question
 
 
 class QuestionForm(forms.ModelForm):
