+++ app/basic_information/users/forms.py
@@ -1,0 +1,8 @@
+from django.contrib.auth.forms import UserCreationForm
+from .models import CustomUser
+
+
+class CustomUserCreationForm(UserCreationForm):
+    class Meta:
+        model = CustomUser
+        fields = ("username", "password1", "password2")  # 必要に応じてフィールド追加
